azFunctions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-eventhub
requests
beautifulsoup4
langchain-openai
langchain-core
openai
langchain
langchain-community
langchain_experimental
